Since python 2.7.9 ssl certificates are rejected by default if invalid and as a result python will object to self-signed certificates (see https://www.python.org/dev/peps/pep-0476/).
The current configuration for Windows clients sets up a self signed cert for for the HTTPS connector. If a 2.7.9 runtime is used to connect to this server then HTTPS connection will be rejected.

This patch allows you to provide a certificate to validate the ssl connection either via the ansible config propert 'ansible_client_validation_cert' or as cli arg --client-validation-cert.